# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "W Craft API",
    "version": "1.0.0",
    "description": "Canonical contract for frontend-facing project, character tree, poster and Music Studio operations."
  },
  "servers": [
    {"url": "/"}
  ],
  "security": [
    {"userToken": []}
  ],
...
  "paths": {
...
    "/api/character/select/": {
      "get": {
        "operationId": "listCharacterTree",
        "parameters": [
          {
            "name": "projectId",
            "in": "query",
            "required": true,
            "schema": {"type": "integer"}
          }
        ],
        "responses": {
          "200": {
            "description": "Character tree",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {"$ref": "#/components/schemas/CharacterTreeNode"}
                }
              }
            }
          },
          "401": {"$ref": "#/components/responses/ApiError"},
          "403": {"$ref": "#/components/responses/ApiError"},
          "404": {"$ref": "#/components/responses/ApiError"}
        }
      }
    },
    "/api/character/create/": {
      "post": {
        "operationId": "createCharacterTreeNode",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {"$ref": "#/components/schemas/CharacterTreeCreateRequest"}
            }
          }
        },
        "responses": {
          "200": {"description": "Tree node created"},
          "400": {"$ref": "#/components/responses/ApiError"},
          "401": {"$ref": "#/components/responses/ApiError"},
          "403": {"$ref": "#/components/responses/ApiError"},
          "404": {"$ref": "#/components/responses/ApiError"}
        }
      }
    },
...
    "/api/character/delete/": {
      "post": {
        "operationId": "deleteCharacterTreeNode",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {"$ref": "#/components/schemas/CharacterTreeDeleteRequest"}
            }
          }
        },
        "responses": {
          "200": {
            "description": "Tree node deleted",
            "content": {
              "application/json": {
                "schema": {"$ref": "#/components/schemas/DeleteResponse"}
              }
            }
          },
          "401": {"$ref": "#/components/responses/ApiError"},
          "403": {"$ref": "#/components/responses/ApiError"},
          "404": {"$ref": "#/components/responses/ApiError"}
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "userToken": {
        "type": "apiKey",
        "in": "header",
        "name": "X-User-Token"
      }
    },
...
    "responses": {
      "ApiError": {
        "description": "Canonical API error",
        "content": {
          "application/json": {
            "schema": {"$ref": "#/components/schemas/ApiErrorEnvelope"}
          }
        }
      },
...
    },
    "schemas": {
      "ApiErrorDetail": {
        "type": "object",
        "required": ["code", "message"],
        "properties": {
          "code": {"type": "string"},
          "message": {"type": "string"},
          "fields": {
            "type": "object",
            "additionalProperties": true
          }
        }
      },
      "ApiErrorEnvelope": {
        "type": "object",
        "required": ["error"],
        "properties": {
          "error": {"$ref": "#/components/schemas/ApiErrorDetail"},
          "code": {"type": "string", "deprecated": true},
          "detail": {"type": "string", "deprecated": true},
          "errors": {"type": "object", "additionalProperties": true, "deprecated": true}
        }
      },
      "CharacterTreeNode": {
        "type": "object",
        "required": ["id", "key", "name", "is_folder"],
        "properties": {
          "id": {"type": "string", "format": "uuid"},
          "key": {"type": "string"},
          "name": {"type": "string"},
          "is_folder": {"type": "boolean"},
          "character_id": {"type": "string", "format": "uuid", "nullable": true},
          "legacy_hero_id": {"type": "integer", "nullable": true},
          "children": {
            "type": "array",
            "items": {"$ref": "#/components/schemas/CharacterTreeNode"}
          }
        }
      },
      "CharacterTreeCreateRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": ["id", "name", "type", "projectId"],
        "properties": {
          "id": {"type": "string", "format": "uuid"},
          "name": {"type": "string", "minLength": 1, "maxLength": 255},
          "type": {"type": "string", "enum": ["leaf", "node"]},
          "projectId": {"oneOf": [{"type": "integer"}, {"type": "string"}]},
          "parent": {"type": "string", "format": "uuid", "nullable": true},
          "heroID": {"type": "integer", "nullable": true},
          "studioCharacterId": {"type": "string", "format": "uuid", "nullable": true}
        }
      },
...
      "CharacterTreeDeleteRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": ["id"],
        "properties": {"id": {"type": "string"}}
      },
      "DeleteResponse": {
        "type": "object",
        "required": ["message"],
        "properties": {"message": {"type": "string"}}
      },
...
    }
  }
}